+++ src/pages/ProhibitorySigns/contexts/PageProvider.tsx
@@ -14,7 +14,7 @@ const ProhibitorySignsPageProvider = ({ children }: Props) => {
   const [activeStepWizard, setActiveStepWizard] = useState(0)
   const [address, setAddress] = useState({} as Address)
   const [queryParams] = useSearchParams()
-  let expertMode = queryParams.get('expertMode') ? true : false
+  const expertMode = queryParams.get('expertMode') ? true : false
   const [showScenarioWizard, setShowScenarioWizard] = useState(true)
   const [vehicle, setVehicle] = useState({} as Vehicle)