+++ src/pages/ProhibitorySigns/hooks/useActiveRoadNetwork.ts
@@ -1,7 +1,8 @@
 import { layerIds } from '../contexts/mapLayersReducer'
 
 export const useActiveRoadNetwork = () => {
-  let activeRoadNetwork: (typeof layerIds)[number] = 'roadNetworkNoRestrictions'
+  const activeRoadNetwork: (typeof layerIds)[number] =
+    'roadNetworkNoRestrictions'
 
   return activeRoadNetwork
 }