+++ test/server-handlers.ts
@@ -1,4 +1,4 @@
-import { rest } from 'msw'
+import { DefaultBodyType, rest } from 'msw'
 
 import { ENDPOINT as ENDPOINT_TRAFFIC_SIGNS } from '../src/api/nationaalwegenbestand/rvv/verkeersborden'
 import { ENDPOINT as ENDPOINT_PROHIBITORY_ROADS } from '../src/api/nationaalwegenbestand/rvv/wegvakken'
@@ -9,15 +9,33 @@ import { ENDPOINT as ENDPOINT_RDW_FUEL } from '../src/api/rdw/fuel'
 import { ENDPOINT as ENDPOINT_RDW_SUBCATEGORY } from '../src/api/rdw/subcategory'
 import { ENDPOINT as ENDPOINT_RDW_VEHICLE } from '../src/api/rdw/vehicle'
 
+// addresses
+import addressDefault from './mocks/pdok/suggest/annas.json'
+import addressNoResults from './mocks/pdok/suggest/no-results.json'
+
+// road sections
+import roadSection600778786 from './mocks/nationaalwegenbestand/wegvakken/600778786.json'
+import roadSectionNotFound from './mocks/nationaalwegenbestand/wegvakken/not-found.json'
+
+// rvv related data
+import rvvRoadSections from './mocks/nationaalwegenbestand/rvv/wegvakken/data.json'
+import rvvTrafficSigns from './mocks/nationaalwegenbestand/rvv/verkeersborden/data.json'
+
+// vehicles
+import onlyTrailer from './mocks/rdw/vehicle/ot77fj.json'
+import vehicleWithTrailer from './mocks/rdw/vehicle/24bjl7.json'
+import vehicleNoMaximumAllowedWeight from './mocks/rdw/vehicle/65jrdp.json'
+import vehicleMobileCrane from './mocks/rdw/vehicle/85bpf2.json'
+import vehicleValidTruck from './mocks/rdw/vehicle/bxls14.json'
+
 export const handlers = [
   rest.get(ENDPOINT_ADDRESS_SEARCH, (req, res, ctx) => {
     const searchResultsMock = getAddressResults(req.url.searchParams)
     return res(ctx.status(200), ctx.json(searchResultsMock))
   }),
 
-  rest.get(ENDPOINT_PROHIBITORY_ROADS, (req, res, ctx) => {
-    const roadsMock = require('./mocks/nationaalwegenbestand/rvv/wegvakken/data.json')
-    return res(ctx.status(200), ctx.json(roadsMock))
+  rest.get(ENDPOINT_PROHIBITORY_ROADS, (_req, res, ctx) => {
+    return res(ctx.status(200), ctx.json(rvvRoadSections))
   }),
 
   rest.get(`/${ENDPOINT_ROAD_SECTION}:roadSectionId`, (req, res, ctx) => {
@@ -26,78 +44,87 @@ export const handlers = [
     return res(ctx.status(200), ctx.json(roadSectionMock))
   }),
 
-  rest.get(ENDPOINT_TRAFFIC_SIGNS, (req, res, ctx) => {
-    const trafficSignsMock = require('./mocks/nationaalwegenbestand/rvv/verkeersborden/data.json')
-    return res(ctx.status(200), ctx.json(trafficSignsMock))
+  rest.get(ENDPOINT_TRAFFIC_SIGNS, (_req, res, ctx) => {
+    return res(ctx.status(200), ctx.json(rvvTrafficSigns))
   }),
 
-  rest.get(ENDPOINT_RDW_AXLES, (req, res, ctx) => {
+  rest.get(ENDPOINT_RDW_AXLES, async (req, res, ctx) => {
     const licensePlate = req.url.searchParams.get('kenteken')
-    const axlesMock = !licensePlate
-      ? []
-      : require(`./mocks/rdw/axles/${licensePlate.toLowerCase()}.json`)
+    let axlesMock: [] | Promise<unknown> = []
+
+    if (licensePlate) {
+      axlesMock = await import(
+        `./mocks/rdw/axles/${licensePlate.toLowerCase()}.json`
+      ).then(module => module.default)
+    }
 
     return res(ctx.status(200), ctx.json(axlesMock))
   }),
 
-  rest.get(ENDPOINT_RDW_FUEL, (req, res, ctx) => {
+  rest.get(ENDPOINT_RDW_FUEL, async (req, res, ctx) => {
     const licensePlate = req.url.searchParams.get('kenteken')
-    const fuelMock = !licensePlate
-      ? []
-      : require(`./mocks/rdw/fuel/${licensePlate.toLowerCase()}.json`)
+    let fuelMock: [] | Promise<unknown> = []
+
+    if (licensePlate) {
+      fuelMock = await import(
+        `./mocks/rdw/fuel/${licensePlate.toLowerCase()}.json`
+      ).then(module => module.default)
+    }
 
     return res(ctx.status(200), ctx.json(fuelMock))
   }),
 
-  rest.get(ENDPOINT_RDW_SUBCATEGORY, (req, res, ctx) => {
+  rest.get(ENDPOINT_RDW_SUBCATEGORY, async (req, res, ctx) => {
     const licensePlate = req.url.searchParams.get('kenteken')
-    const subcategoryMock = !licensePlate
-      ? []
-      : require(`./mocks/rdw/subcategory/${licensePlate.toLowerCase()}.json`)
+    let subcategoryMock: [] | Promise<unknown> = []
+
+    if (licensePlate) {
+      subcategoryMock = await import(
+        `./mocks/rdw/subcategory/${licensePlate.toLowerCase()}.json`
+      ).then(module => module.default)
+    }
 
     return res(ctx.status(200), ctx.json(subcategoryMock))
   }),
 
-  rest.get(ENDPOINT_RDW_VEHICLE, (req, res, ctx) => {
-    const vehicleMock = getVehicle(req.url.searchParams)
+  rest.get(ENDPOINT_RDW_VEHICLE, async (req, res, ctx) => {
+    const vehicleMock = await getVehicle(req.url.searchParams)
     return res(ctx.status(vehicleMock.status), ctx.json(vehicleMock.body))
   }),
 ]
 
 const getAddressResults = (params: URLSearchParams) => {
-  let result = require('./mocks/pdok/suggest/annas.json')
-
   if (params.get('q')?.includes('Noresults')) {
-    result = require('./mocks/pdok/suggest/no-results.json')
+    return addressNoResults
   }
 
   if (params.get('q')?.includes('API500')) {
-    result = { status: 500, body: {} }
+    return { status: 500, body: {} }
   }
 
-  return result
+  return addressDefault
 }
 
 const getRoadSection = (id: string | ReadonlyArray<string>) => {
   if (id === '600778786') {
-    return require('./mocks/nationaalwegenbestand/wegvakken/600778786.json')
+    return roadSection600778786
   }
 
   if (id === '404404') {
-    return require('./mocks/nationaalwegenbestand/wegvakken/not-found.json')
+    return roadSectionNotFound
   }
 
-  return console.error('no roadSection mock found.')
+  return roadSectionNotFound
 }
 
-const getVehicle = (params: URLSearchParams) => {
+const getVehicle = async (params: URLSearchParams) => {
   const licensePlate = params.get('kenteken')
-  const mocks: Record<string, string> = {
-    '24BJL7': require('./mocks/rdw/vehicle/24bjl7.json'), // vehicle + trailer use case
-    '65JRDP': require('./mocks/rdw/vehicle/65jrdp.json'), // no maximum allowed weight
-    '85BPF2': require('./mocks/rdw/vehicle/85bpf2.json'), // mobile crane
-    BXLS14: require('./mocks/rdw/vehicle/bxls14.json'), // valid vehicle
-    OT77FJ: require('./mocks/rdw/vehicle/ot77fj.json'), // trailer
+  const mocks: Record<string, DefaultBodyType> = {
+    '24BJL7': vehicleWithTrailer,
+    '65JRDP': vehicleNoMaximumAllowedWeight,
+    '85BPF2': vehicleMobileCrane,
+    BXLS14: vehicleValidTruck,
+    OT77FJ: onlyTrailer,
   }
 
   if (licensePlate === 'API429') {
@@ -108,7 +135,10 @@ const getVehicle = (params: URLSearchParams) => {
     return { status: 500, body: {} }
   }
 
-  if (licensePlate && Object.hasOwn(mocks, licensePlate)) {
+  if (
+    licensePlate &&
+    Object.prototype.hasOwnProperty.call(mocks, licensePlate)
+  ) {
     return { status: 200, body: mocks[licensePlate] }
   }